LLM+Evaluation/requirements.txt
deepeval
python-dotenv>=1.0.0
pydantic_settings